+++ src/benchmarks/real-world/bepuphysics2/Program.cs
@@ -8,4 +8,3 @@ public class BepuPhysics2Benchmarks
         BenchmarkSwitcher.FromAssembly(typeof(BepuPhysics2Benchmarks).Assembly).Run(args);
     }
 }
-
