+++ src/benchmarks/real-world/bepuphysics2/SweepBenchmarks.cs
@@ -12,12 +12,9 @@ using static DemoBenchmarks.BenchmarkHelper;
 
 namespace DemoBenchmarks;
 
-/// <summary>
-/// Evaluates performance of all sweep tests.
-/// </summary>
-public class Sweeps
+public class Sweeper
 {
-    const int iterationCount = 100;
+    const int iterationCount = 10;
     BufferPool pool;
     struct Pair
     {
@@ -34,9 +31,7 @@ public class Sweeps
     SweepTaskRegistry taskRegistry;
     Shapes shapes;
 
-
-    [GlobalSetup]
-    public unsafe void Setup()
+    public Sweeper()
     {
         pool = new BufferPool();
         pool.Take(iterationCount, out pairs);
@@ -109,7 +104,6 @@ public class Sweeps
         }
     }
 
-    [GlobalCleanup]
     public void Cleanup()
     {
         //All outstanding allocations poof when the pool is cleared.
@@ -125,7 +119,7 @@ public class Sweeps
         }
     }
 
-    unsafe Vector3 Test<TA, TB>() where TA : unmanaged, IShape where TB : unmanaged, IShape
+    public unsafe Vector3 Test<TA, TB>() where TA : unmanaged, IShape where TB : unmanaged, IShape
     {
         var task = taskRegistry.GetTask<TA, TB>();
         var aType = default(TA).TypeId;
@@ -146,94 +140,119 @@ public class Sweeps
         }
         return resultSum;
     }
+}
 
+/// <summary>
+/// Evaluates performance of a representative subset of sweep tests.
+/// </summary>
+public class SweepBenchmarks
+{
+    Sweeper sweeper;
+    [GlobalSetup]
+    public unsafe void Setup()
+    {
+        sweeper = new Sweeper();
+    }
+
+    [GlobalCleanup]
+    public void Cleanup()
+    {
+        sweeper.Cleanup();
+    }
+
+
+    //Some (commented) benchmarks are punted to SweepBenchmarksDeep:
+    //-Convex-compound tests are pretty similar to each other. Box-compound/bigcompound/mesh is preserved, but the rest are commented. Likewise, bigcompound-bigcompound is the only compound-compound pair that's tested.
+    //-Multiple sweeps use GJK for the distance tester. While the support functions used vary, those are relatively low value compared to the GJK outer loop. If we really need to, we can cover the support functions separately.
+    //Most sphere tests aren't super interesting, so we exclude some of them.
+
+    //[Benchmark]
+    //public void SphereSphere() => sweeper.Test<Sphere, Sphere>();
     [Benchmark]
-    public void SphereSphere() => Test<Sphere, Sphere>();
+    public void SphereCapsule() => sweeper.Test<Sphere, Capsule>();
+    //[Benchmark]
+    //public void SphereBox() => sweeper.Test<Sphere, Box>();
     [Benchmark]
-    public void SphereCapsule() => Test<Sphere, Capsule>();
+    public void SphereTriangle() => sweeper.Test<Sphere, Triangle>();
+    //[Benchmark]
+    //public void SphereCylinder() => sweeper.Test<Sphere, Cylinder>();
+    //[Benchmark]
+    //public void SphereConvexHull() => Test<Sphere, ConvexHull>(); //GJK
+    //[Benchmark]
+    //public void SphereCompound() => Test<Sphere, Compound>();
+    //[Benchmark]
+    //public void SphereBigCompound() => Test<Sphere, BigCompound>();
+    //[Benchmark]
+    //public void SphereMesh() => Test<Sphere, Mesh>();
     [Benchmark]
-    public void SphereBox() => Test<Sphere, Box>();
+    public void CapsuleCapsule() => sweeper.Test<Capsule, Capsule>();
     [Benchmark]
-    public void SphereTriangle() => Test<Sphere, Triangle>();
+    public void CapsuleBox() => sweeper.Test<Capsule, Box>();
+    //[Benchmark]
+    //public void CapsuleTriangle() => Test<Capsule, Triangle>(); //GJK
     [Benchmark]
-    public void SphereCylinder() => Test<Sphere, Cylinder>();
+    public void CapsuleCylinder() => sweeper.Test<Capsule, Cylinder>(); //GJK
+    //[Benchmark]
+    //public void CapsuleConvexHull() => Test<Capsule, ConvexHull>(); //GJK
+    //[Benchmark]
+    //public void CapsuleCompound() => Test<Capsule, Compound>();
+    //[Benchmark]
+    //public void CapsuleBigCompound() => Test<Capsule, BigCompound>();
+    //[Benchmark]
+    //public void CapsuleMesh() => Test<Capsule, Mesh>();
+    //[Benchmark]
+    //public void BoxBox() => Test<Box, Box>(); //GJK
+    //[Benchmark]
+    //public void BoxTriangle() => Test<Box, Triangle>(); //GJK
+    //[Benchmark]
+    //public void BoxCylinder() => Test<Box, Cylinder>(); //GJK
+    //[Benchmark]
+    //public void BoxConvexHull() => Test<Box, ConvexHull>(); //GJK
     [Benchmark]
-    public void SphereConvexHull() => Test<Sphere, ConvexHull>();
+    public void BoxCompound() => sweeper.Test<Box, Compound>();
+    //[Benchmark]
+    //public void BoxBigCompound() => Test<Box, BigCompound>(); //Well covered by bigcompound-bigcompound.
     [Benchmark]
-    public void SphereCompound() => Test<Sphere, Compound>();
+    public void BoxMesh() => sweeper.Test<Box, Mesh>();
+    //[Benchmark]
+    //public void TriangleTriangle() => Test<Triangle, Triangle>(); //GJK
+    //[Benchmark]
+    //public void TriangleCylinder() => Test<Triangle, Cylinder>(); //GJK
+    //[Benchmark]
+    //public void TriangleConvexHull() => Test<Triangle, ConvexHull>(); //GJK
+    //[Benchmark]
+    //public void TriangleCompound() => Test<Triangle, Compound>();
+    //[Benchmark]
+    //public void TriangleBigCompound() => Test<Triangle, BigCompound>();
+    //[Benchmark]
+    //public void TriangleMesh() => Test<Triangle, Mesh>();
+    //[Benchmark]
+    //public void CylinderCylinder() => Test<Cylinder, Cylinder>(); //GJK
+    //[Benchmark]
+    //public void CylinderConvexHull() => Test<Cylinder, ConvexHull>(); //GJK
+    //[Benchmark]
+    //public void CylinderCompound() => Test<Cylinder, Compound>();
+    //[Benchmark]
+    //public void CylinderBigCompound() => Test<Cylinder, BigCompound>();
+    //[Benchmark]
+    //public void CylinderMesh() => Test<Cylinder, Mesh>();
+    //[Benchmark]
+    //public void ConvexHullConvexHull() => Test<ConvexHull, ConvexHull>(); //GJK
+    //[Benchmark]
+    //public void ConvexHullCompound() => Test<ConvexHull, Compound>();
+    //[Benchmark]
+    //public void ConvexHullBigCompound() => Test<ConvexHull, BigCompound>();
+    //[Benchmark]
+    //public void ConvexHullMesh() => Test<ConvexHull, Mesh>();
+    //[Benchmark]
+    //public void CompoundCompound() => Test<Compound, Compound>();
+    //[Benchmark]
+    //public void CompoundBigCompound() => Test<Compound, BigCompound>();
+    //[Benchmark]
+    //public void CompoundMesh() => Test<Compound, Mesh>();
     [Benchmark]
-    public void SphereBigCompound() => Test<Sphere, BigCompound>();
-    [Benchmark]
-    public void SphereMesh() => Test<Sphere, Mesh>();
-    [Benchmark]
-    public void CapsuleCapsule() => Test<Capsule, Capsule>();
-    [Benchmark]
-    public void CapsuleBox() => Test<Capsule, Box>();
-    [Benchmark]
-    public void CapsuleTriangle() => Test<Capsule, Triangle>();
-    [Benchmark]
-    public void CapsuleCylinder() => Test<Capsule, Cylinder>();
-    [Benchmark]
-    public void CapsuleConvexHull() => Test<Capsule, ConvexHull>();
-    [Benchmark]
-    public void CapsuleCompound() => Test<Capsule, Compound>();
-    [Benchmark]
-    public void CapsuleBigCompound() => Test<Capsule, BigCompound>();
-    [Benchmark]
-    public void CapsuleMesh() => Test<Capsule, Mesh>();
-    [Benchmark]
-    public void BoxBox() => Test<Box, Box>();
-    [Benchmark]
-    public void BoxTriangle() => Test<Box, Triangle>();
-    [Benchmark]
-    public void BoxCylinder() => Test<Box, Cylinder>();
-    [Benchmark]
-    public void BoxConvexHull() => Test<Box, ConvexHull>();
-    [Benchmark]
-    public void BoxCompound() => Test<Box, Compound>();
-    [Benchmark]
-    public void BoxBigCompound() => Test<Box, BigCompound>();
-    [Benchmark]
-    public void BoxMesh() => Test<Box, Mesh>();
-    [Benchmark]
-    public void TriangleTriangle() => Test<Triangle, Triangle>();
-    [Benchmark]
-    public void TriangleCylinder() => Test<Triangle, Cylinder>();
-    [Benchmark]
-    public void TriangleConvexHull() => Test<Triangle, ConvexHull>();
-    [Benchmark]
-    public void TriangleCompound() => Test<Triangle, Compound>();
-    [Benchmark]
-    public void TriangleBigCompound() => Test<Triangle, BigCompound>();
-    [Benchmark]
-    public void TriangleMesh() => Test<Triangle, Mesh>();
-    [Benchmark]
-    public void CylinderCylinder() => Test<Cylinder, Cylinder>();
-    [Benchmark]
-    public void CylinderConvexHull() => Test<Cylinder, ConvexHull>();
-    [Benchmark]
-    public void CylinderCompound() => Test<Cylinder, Compound>();
-    [Benchmark]
-    public void CylinderBigCompound() => Test<Cylinder, BigCompound>();
-    [Benchmark]
-    public void CylinderMesh() => Test<Cylinder, Mesh>();
-    [Benchmark]
-    public void ConvexHullConvexHull() => Test<ConvexHull, ConvexHull>();
-    [Benchmark]
-    public void ConvexHullCompound() => Test<ConvexHull, Compound>();
-    [Benchmark]
-    public void ConvexHullBigCompound() => Test<ConvexHull, BigCompound>();
-    [Benchmark]
-    public void ConvexHullMesh() => Test<ConvexHull, Mesh>();
-    [Benchmark]
-    public void CompoundCompound() => Test<Compound, Compound>();
-    [Benchmark]
-    public void CompoundBigCompound() => Test<Compound, BigCompound>();
-    [Benchmark]
-    public void CompoundMesh() => Test<Compound, Mesh>();
-    [Benchmark]
-    public void BigCompoundBigCompound() => Test<BigCompound, BigCompound>();
-    [Benchmark]
-    public void BigCompoundMesh() => Test<BigCompound, Mesh>();
+    public void BigCompoundBigCompound() => sweeper.Test<BigCompound, BigCompound>();
+    //[Benchmark]
+    //public void BigCompoundMesh() => Test<BigCompound, Mesh>();
     //No mesh-mesh!
 }
